Feat/dmau 559/shop entry api change (#50)
* implemented the Plastic Bag Fee feature + redesign the server home page

* set cors to be true & remove Hoboken of 7EF2370

* remove Hoboken of 7EF2370

* chore(config): add YAML Substitute for cors

Added a yaml substitution for cors. This lays the groundwork for a dev
and production config that allows us to deploy the same way to multiple
environments.

To use this, use the following flag:

```sh
c

* wip

* chore(docs): change caution to warning

* enhance the README.md with more content for the Blue-Green deployment steps, and add some more contents

* chore(docs): split internal and external services

* chore(docs): add rolling deployment link

V3 of Cloud Foundry will have a [rolling release](https://docs.cloudfoundry.org/devguide/deploy-apps/rolling-deploy.html) so we do not need setup secondary servers manually. May be worthwhile to begin experimenting with this in development.

* correct some typo

* remove an extra variable of cors in manifest.yml and update some typo in README.md

* added auto scaling & papertrail logging service to manifest files

* added patch to route shop entry to correct one in v7

* swapped mobileapi for api for shop entry call